test: fix the integrations for subscriptions
This is a bug fix for integration test about subscriptions. Initially we
thought the response from FxA account API would only update subscription
status if the subscription of the user expired. It turned out that the
whole schema of the response changed. We adjust our mock API server
based on the changes.